+++ src/main/java/com/showroommanagement/service/ShowroomService.java
@@ -1,45 +1,54 @@
 package com.showroommanagement.service;
 
+import com.showroommanagement.dto.ResponseDTO;
 import com.showroommanagement.entity.Showroom;
+import com.showroommanagement.exception.BadRequestServiceAlertException;
 import com.showroommanagement.repository.ShowroomRepository;
+import com.showroommanagement.util.Constant;
+import jakarta.transaction.Transactional;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
-
-import java.util.List;
-import java.util.NoSuchElementException;
-import java.util.Optional;
 
 @Service
 public class ShowroomService {
     private final ShowroomRepository showroomRepository;
 
-    public ShowroomService(final ShowroomRepository showroomRepository) {
+    public ShowroomService(ShowroomRepository showroomRepository) {
         this.showroomRepository = showroomRepository;
     }
 
-    public Showroom createShowroom(final Showroom showroom) {
-        return this.showroomRepository.save(showroom);
+    @Transactional
+    public ResponseDTO createShowroom(final Showroom showroom) {
+        final ResponseDTO responseDTO = new ResponseDTO();
+        responseDTO.setMessage(Constant.CREATE);
+        responseDTO.setStatusCode(HttpStatus.CREATED.value());
+        responseDTO.setData(this.showroomRepository.save(showroom));
+        return responseDTO;
     }
 
-    public Showroom retrieveById(final Integer id) {
-        return this.showroomRepository.findById(id).orElseThrow(() -> new NoSuchElementException("No Showroom found for this id"));
-    }
-
-    public List<Showroom> retrieveAll() {
-        return this.showroomRepository.findAll();
-    }
-
-
-//    public List <BikeDetail> retrieveBySearch(final String search){
-//        return this.showroomRepository.retrieveBySearch(search);
-//
-//    }
-
-    public Showroom updateById(final Showroom showroom, final Integer id) {
-        final Optional<Showroom> showroomOptional = this.showroomRepository.findById(id);
-        if (showroomOptional.isEmpty()) {
-            throw new IllegalArgumentException("Showroom Not Found");
+    public ResponseDTO retrieveById(final Integer id) {
+        if (this.showroomRepository.existsById(id)) {
+            final ResponseDTO responseDTO = new ResponseDTO();
+            responseDTO.setMessage(Constant.RETRIEVE);
+            responseDTO.setStatusCode(HttpStatus.OK.value());
+            responseDTO.setData(this.showroomRepository.findById(id));
+            return responseDTO;
+        } else {
+            throw new BadRequestServiceAlertException(Constant.ID_DOES_NOT_EXIST);
         }
-        final Showroom showroomObject = showroomOptional.get();
+    }
+
+    public ResponseDTO retrieveAll() {
+        final ResponseDTO responseDTO = new ResponseDTO();
+        responseDTO.setMessage(Constant.RETRIEVE);
+        responseDTO.setStatusCode(HttpStatus.OK.value());
+        responseDTO.setData(this.showroomRepository.findAll());
+        return responseDTO;
+    }
+
+    @Transactional
+    public ResponseDTO updateById(final Showroom showroom, final Integer id) {
+        final Showroom showroomObject = this.showroomRepository.findById(id).orElseThrow(() -> new BadRequestServiceAlertException(Constant.ID_DOES_NOT_EXIST));
         if (showroom.getName() != null) {
             showroomObject.setName(showroom.getName());
         }
@@ -52,16 +61,26 @@ public class ShowroomService {
         if (showroom.getContactNumber() != 0) {
             showroomObject.setContactNumber(showroom.getContactNumber());
         }
-        return this.showroomRepository.save(showroomObject);
+        final ResponseDTO responseDTO = new ResponseDTO();
+        responseDTO.setMessage(Constant.UPDATE);
+        responseDTO.setStatusCode(HttpStatus.OK.value());
+        responseDTO.setData(this.showroomRepository.save(showroomObject));
+        return responseDTO;
     }
 
-    public String deleteById(final Integer id) {
+    public ResponseDTO deleteById(final Integer id) {
         if (id == null) {
-            throw new IllegalArgumentException("Showroom ID cannot be null.");
+            throw new BadRequestServiceAlertException(Constant.DATA_NULL);
         }
-        final Showroom showroom = this.showroomRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("Showroom Id Not found"));
-        this.showroomRepository.deleteById(id);
-        return "Deleted Successfully";
+        if (this.showroomRepository.existsById(id)) {
+            this.showroomRepository.deleteById(id);
+            final ResponseDTO responseDTO = new ResponseDTO();
+            responseDTO.setMessage(Constant.DELETE);
+            responseDTO.setStatusCode(HttpStatus.OK.value());
+            responseDTO.setData(Constant.REMOVE);
+            return responseDTO;
+        } else {
+            throw new BadRequestServiceAlertException(Constant.ID_DOES_NOT_EXIST);
+        }
     }
-
 }
